Answer some comments from AB+PK
