+++ scilpy/io/gradients.py
