+++ scripts/scil_extract_dwi_shell.py
@@ -25,7 +25,7 @@ import nibabel as nib
 import numpy as np
 
 from scilpy.dwi.utils import extract_dwi_shell
-from scilpy.io.gradient_table import save_gradient_sampling_fsl
+from scilpy.io.gradients import save_gradient_sampling_fsl
 from scilpy.io.utils import (add_overwrite_arg, add_verbose_arg,
                              assert_inputs_exist, assert_outputs_exist)
