+++ scripts/scil_flip_or_swap_gradients_axis.py
@@ -10,10 +10,11 @@ import os
 
 import numpy as np
 
-from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
-                             assert_outputs_exist)
 from scilpy.gradients.bvec_bval_tools import (flip_gradient_sampling,
                                               swap_gradient_axis)
+from scilpy.gradients.utils import str_to_flip_swap
+from scilpy.io.utils import (add_overwrite_arg, assert_inputs_exist,
+                             assert_outputs_exist)
 from scilpy.utils.util import str_to_index
